+++ image-processor-stl-5/src/main/java/imageManipulation/ImageDisplayWindow.java
@@ -13,7 +13,7 @@ public class ImageDisplayWindow extends JFrame {
     
     private final BufferedImage image;
 
-    public ImageDisplayWindow(int[][][] imageArray, String functionName, int seqNum, int sourceSeqNum, String fileName) {
+    public ImageDisplayWindow(byte[][][] imageArray, String functionName, int seqNum, int sourceSeqNum, String fileName) {
         super(buildTitle(functionName, seqNum, sourceSeqNum, imageArray, fileName));
         
         this.image = ImageConverter.arrayToBufferedImage(imageArray);
@@ -29,7 +29,7 @@ public class ImageDisplayWindow extends JFrame {
         setLocationRelativeTo(null);
     }
     
-    private static String buildTitle(String functionName, int seqNum, int sourceSeqNum, int[][][] imageArray, String fileName) {
+    private static String buildTitle(String functionName, int seqNum, int sourceSeqNum, byte[][][] imageArray, String fileName) {
         int width = imageArray[0].length;
         int height = imageArray.length;
